backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
httpx==0.25.1
python-dotenv==1.0.0
supabase==1.2.0
asyncpg==0.29.0
redis==5.0.1
python-multipart==0.0.6
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
psycopg2-binary==2.9.9